#1: Update models.py remove extra meta
Hiya Kevin,  

So I've decided to get my github account giving me half-decent signals and I am up to here. 

This issue has been open for 17 years.

There is no question that this project has been pining for the fjords for some time, but it's an easy change and coders gotta code. I'm using my best judgement about the `sort_order`.

I understand it's now on you to have the day I'm having, but look, it's not so bad.

Hoping this is finding you well.